#315 rearrange tests and update names
